move to shared types and shared components
