+++ components/shared/ViewTypePill.vue
