+++ types/index.ts
@@ -125,6 +125,24 @@ export interface Views {
 
 export type ViewType = "alerts" | "map" | "gallery";
 
+export const VIEW_TYPES = [
+  "alerts",
+  "map",
+  "gallery",
+] as const satisfies readonly ViewType[];
+
+export const SECONDARY_DATASET_VIEW_TYPES = [
+  "alerts",
+  "map",
+] as const satisfies readonly ViewType[];
+
+/** Whether a view type can reference a companion warehouse table. */
+export const supportsSecondaryDataset = (
+  viewType: ViewType | undefined,
+): boolean =>
+  viewType !== undefined &&
+  SECONDARY_DATASET_VIEW_TYPES.some((type) => type === viewType);
+
 export type AllowedFileExtensions = {
   audio: string[];
   image: string[];
